moved shuffle to helpers
