+++ lib/helpers.js
@@ -47,3 +47,23 @@ var validKeyType = exports.validKeyType = function(mockInstance, key, type, call
 var initKey = exports.initKey = function(mockInstance, key, fn) {
   mockInstance.storage[key] = mockInstance.storage[key] || fn();
 };
+
+var shuffle = exports.shuffle = function(array) {
+  var counter = array.length;
+
+  // While there are elements in the array
+  while (counter > 0) {
+    // Pick a random index
+    var index = Math.floor(Math.random() * counter);
+
+    // Decrease counter by 1
+    counter--;
+
+    // And swap the last element with it
+    var temp = array[counter];
+    array[counter] = array[index];
+    array[index] = temp;
+  }
+
+  return array;
+};
