+++ lib/set.js
@@ -1,4 +1,5 @@
 var Item = require('./item.js');
+var shuffle = require('./helpers.js').shuffle;
 
 /**
  * Sadd
@@ -177,25 +178,5 @@ exports.srandmember = function (mockInstance, key, count, callback) {
   }
 
   mockInstance._callCallback(callback, null, result);
-}
-
-function shuffle(array) {
-  var counter = array.length;
-
-  // While there are elements in the array
-  while (counter > 0) {
-    // Pick a random index
-    var index = Math.floor(Math.random() * counter);
-
-    // Decrease counter by 1
-    counter--;
-
-    // And swap the last element with it
-    var temp = array[counter];
-    array[counter] = array[index];
-    array[index] = temp;
-  }
-
-  return array;
 }
